Fix a number of compile time warnings so that uClibc will build with -Werror using a 3.0.4 version of the sh-linux-gcc compiler.

+++ include/features.h
@@ -422,8 +422,8 @@ uClibc was built without large file support enabled.
 #define __builtin_expect(x, expected_value) (x)
 #endif
 
-#define likely(x)       __builtin_expect((x),1)
-#define unlikely(x)     __builtin_expect((x),0)
+#define likely(x)       __builtin_expect(((int)(x)),1)
+#define unlikely(x)     __builtin_expect(((int)(x)),0)
 
 /* --- this is added to integrate linuxthreads */
 #define __USE_UNIX98            1
